+++ CHANGELOG.md
@@ -19,7 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(None)
+### Added
+
+- Added overload for AttemptsTo that accepts multiple tasks
+
+### Changed
+
+- Updated AbstractComparison constructor to protected
 
 
 ## [1.0.0] - 2021-05-24